% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_plot}
\alias{bar_plot}
\title{bar_plot}
\usage{
bar_plot(plotting_data, main = "", xlab = "", ylab = "\%", las = 1,
  dataset_names = colnames(plotting_data), order_numerically = FALSE,
  ordered_categories = NULL, plot_differences = FALSE,
  cex_y_axis = par("cex.axis"), cex_names = par("cex.axis"),
  col = NULL, legend = TRUE, legend_cex = 1)
}
\arguments{
\item{plotting_data}{dataframe or vector containing numerical data}

\item{main}{title}

\item{dataset_names}{vector of names of the datasets in the plotting_data list,
by default these are taken from the column names of the dataframe.}
}
\value{
produces a barplot and returns values generated by base function barplot.
}
\description{
barplot to compare 2 samples
}
\details{
Intended for use with chromosome and (biotype - not sure if we're using this here) data.
Uses the base function barplot and passes extra parameters to it.
}
\examples{
bar_plot(chromosomes)
bar_plot(chromosomes, cex_names = 0.8, col = topo.colors(2), ylab = "\% of genes", xlab = "chromosome")
}
