% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chromosomes}
\alias{chromosomes}
\title{Set of chromosome data}
\format{A data frame containing 2 columns. The row names are the chromosome names}
\usage{
chromosomes
}
\description{
A dataset derived from 2 sets of genes, containing the proportion of genes that
were located on each chromosome. The query set are \code{genes_1} and the
background set \code{bg_genes}.
}
\keyword{datasets}
