% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(plotting_data, main = "", xlab = "", log = FALSE,
  dataset_names = names(plotting_data), legend = TRUE,
  legend_cex = 1, col = NULL, legend_pos = "topright")
}
\arguments{
\item{plotting_data}{list of numerical vectors}

\item{main}{title}

\item{log}{whether to log2 transform the data \link{default: false}}

\item{dataset_names}{vector of names of the datasets in the plotting_data list, by default these are
taken from the names of the components in the list}
}
\value{
produces a density plot.
}
\description{
Density (line) plot of multiple samples
}
\details{
Intended for use with GC content and gene length data
}
\examples{
density_plot(some_data)
}
