% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{extract_gene_names}
\alias{extract_gene_names}
\title{Extract gene names from GTF}
\usage{
extract_gene_names(info_column)
}
\arguments{
\item{info_column}{character vector from the info column of a gtf file, probably the output from \code{\link{process_GTF}}, in the
format "gene_id "ENSMUSG00000102693"; gene_version "1"; gene_name "4933401J01Rik"; gene_source "havana"; gene_biotype "TEC"}
}
\value{
A data frame containing the gene sets of interest along with information on
gene lengths etc. \url{http://rstudio.com} some more text \link{process_GMT}
}
\description{
Parse info column from GTF file to extract gene names
}
\examples{
x <- extract_gene_names(m_musculus$info)
}
