% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_chr_percentage}
\alias{get_chr_percentage}
\title{get_chr_percentage}
\usage{
get_chr_percentage(chr_list)
}
\arguments{
\item{chr_list}{named list containing character vectors of chromosome names}
}
\description{
takes a named list (e.g. query and background) containing vectors of
chromosome names
}
