% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_chromosomes}
\alias{get_chromosomes}
\title{get_chromosomes.}
\usage{
get_chromosomes(query_names, gene_info, name_column = NULL,
  chr_column = NULL)
}
\arguments{
\item{query_names}{character vector of gene names}

\item{gene_info}{data frame, probably generated by \code{\link{process_GMT}},
containing genes and associated information including location of genes}
}
\value{
A data frame containing the gene sets of interest along with information on
chrs etc. \url{http://rstudio.com} some more text \link{process_GMT}
}
\description{
Looks up gene names from a dataframe and returns the chromosome on which the gene is located
}
\examples{
get_chr()
}
\keyword{internal}
