% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_count_matrix}
\alias{get_count_matrix}
\title{get_count_matrix.}
\usage{
get_count_matrix(x, name_column = NULL, chr_column = NULL)
}
\arguments{
\item{query_names}{character vector of gene names}

\item{gene_info}{data frame, probably generated by \code{\link{process_GMT}},
containing genes and associated information including location of genes}
}
\value{
A data frame containing the gene sets of interest along with information on
chrs etc. \url{http://rstudio.com} some more text \link{process_GMT}
}
\description{
create a count matrix from a list
}
\examples{
get_chr()
}
\keyword{internal}
