% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{get_GO_info}
\alias{get_GO_info}
\title{Get information on selected gene ontology categories.}
\usage{
get_GO_info(selected_go_names, all_go_categories, gtf)
}
\arguments{
\item{selected_go_names}{names of gene sets of interest}

\item{all_go_categories}{gene sets that have been imported from \code{\link{process_GMT}}}

\item{gtf}{dataframe containing genes and information, probably the output from \code{\link{process_GTF}}}
}
\value{
A data frame containing the gene sets of interest along with information on
gene lengths etc. \url{http://rstudio.com} some more text \link{process_GMT}
}
\description{
Uses information from gtf file to get lengths of genes etc.
}
\examples{
get_GO_info(selected_go_names, all_go_categories, gtf)
}
