% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{get_GO_sets}
\alias{get_GO_sets}
\title{get_GO_sets.}
\usage{
get_GO_sets(selected_go_names, all_go_categories)
}
\arguments{
\item{selected_go_names}{names of gene sets of interest}

\item{all_go_categories}{gene sets that have been imported from \code{\link{process_GMT}}}

\item{gtf}{dataframe containing genes and information, probably the output from \code{\link{process_GTF}}}
}
\description{
gets sets of genes
}
