% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_lengths}
\alias{get_lengths}
\title{Get lengths of genes.}
\usage{
get_lengths(query_names, gene_info, name_column = NULL,
  length_column = NULL)
}
\arguments{
\item{query_names}{character vector of gene names}

\item{gene_info}{data frame, probably generated by \code{\link{process_GMT}},
containing genes and associated information including length of genes}
}
\value{
A data frame containing the gene sets of interest along with information on
gene lengths etc. \url{http://rstudio.com} some more text \link{process_GMT}
}
\description{
Looks up gene names from a dataframe and returns the lengths
}
\examples{
get_lengths()
}
\keyword{internal}
