% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{import_GTF}
\alias{import_GTF}
\title{Import GTF file.}
\usage{
import_GTF(file, genes_only = TRUE)
}
\arguments{
\item{file}{file path for gtf file}

\item{genes_only}{if true, only return the entries annotated as "gene"}
}
\value{
A data frame containing all the genes from \code{gtf_file}
}
\description{
Import GTF file.
}
\examples{
import_GTF("~/file.gtf")
}
