% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{m_musculus}
\alias{m_musculus}
\title{Mouse genome annotations}
\format{A data frame containing information on the Mus musculus genome.}
\source{
\url{ftp://ftp.ensembl.org/pub/release-94/gtf/mus_musculus/Mus_musculus.GRCm38.94.gtf.gz}
}
\usage{
m_musculus
}
\description{
A dataset containing gene annotations from Ensembl for the mouse genome.
}
\details{
This is a gtf file from Ensembl that has been imported into R and saved as a dataframe.
Each line is an annotation for a region in the genome.
The data has been filtered to only include the gene annotations. The original
file also included transcripts, coding sequences, 3' and 5' UTRs etc.
}
\examples{
head(m_musculus)
}
\keyword{datasets}
