% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{parse_GTF_info}
\alias{parse_GTF_info}
\title{Process GTF file.}
\usage{
parse_GTF_info(gtf, gene_names = TRUE, gene_length = TRUE,
  remove_duplicates = TRUE)
}
\arguments{
\item{gtf}{gtf file that has been imported using \code{import_GTF}.
The gtf file must include a column named info that contains data in the following format: a character vector in the
format "gene_id "ENSMUSG00000102693"; gene_version "1"; gene_name "4933401J01Rik"; gene_source "havana"; gene_biotype "TEC"}

\item{gene_names}{if true, add a column containing gene name}

\item{gene_length}{if true, add a column containing gene length}

\item{remove_duplicates}{remove rows if gene names are duplicated. The row that is kept
is the first instance of that gene. To remove duplicates based on something other that
gene name, use the separate \code{remove_duplicates} function.}
}
\value{
A data frame containing all the genes from \code{gtf_file}
}
\description{
parses the info column of a GTF file to extract the gene name and length.
}
\examples{
parse_GTF_info(gtf_file)
}
