% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.r
\name{process_GMT}
\alias{process_GMT}
\title{Import and process GMT file.}
\usage{
process_GMT(file, min_genes = 3, max_genes = 5000)
}
\arguments{
\item{file}{file path for gmt file containing gene sets}

\item{min_genes}{minimum number of genes required in the functional category for the category to be loaded.}

\item{max_genes}{maximum number of genes required in the functional category for the category to be loaded.}
}
\value{
A named list containing all the gene sets imported from \code{gmt_file_path}
}
\description{
Import GMT file containing gene sets.
}
\details{
The GMT format is described \href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:\emph{Gene_Matrix_Transposed_file_format}.28.2A.gmt.29}{GMT format}.
The first column should contain gene set names, the 2nd column contains a brief description and the 3rd column contains a whitespace separated list of genes.
}
\examples{
file <- "http://download.baderlab.org/EM_Genesets/current_release/Mouse/symbol/Mouse_GO_AllPathways_no_GO_iea_December_01_2018_symbol.gmt"
x <- process_GMT(file, 5, 5000)
}
