% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{suspects1}
\alias{suspects1}
\title{Gene set functional categories.}
\format{A named list with x number of components}
\source{
http://download.baderlab.org/EM_Genesets/current_release/Mouse/symbol/Mouse_GO_AllPathways_no_GO_iea_February_01_2019_symbol.gmt

http://download.baderlab.org/EM_Genesets/current_release/Mouse/symbol/Mouse_GO_AllPathways_no_GO_iea_February_01_2019_symbol.gmt
}
\usage{
suspects1
}
\description{
A dataset containing genes grouped in to functional sets. Each functional
group is named. This dataset is processed from a gmt file.
Filtering for reasonable sized categories has been turned off. This set of
categories should mainly be used for testing for biases.
}
\details{
all_go_categories <- process_GMT(file_location, min_genes = 3, max_genes = 500000)
Gene set functional categories.

A dataset containing genes grouped in to functional sets. Each functional
group is named. This dataset is processed from a gmt file.
The

all_go_categories <- process_GMT(file_location)
Set of suspect gene ontology categories

These categories are derived from an analysis of public datasets ....
Essentially these GO categories appeared more than expected when running GO
overrepresentation tests
}
\keyword{datasets}
